% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date2EpochMilli.R
\name{Date2EpochMilli}
\alias{Date2EpochMilli}
\alias{Date2EpochMilli.POSIXt}
\alias{Date2EpochMilli.numeric}
\title{Convert a date into an epoch in milliseconds}
\usage{
Date2EpochMilli(ts)

\method{Date2EpochMilli}{POSIXt}(ts)

\method{Date2EpochMilli}{numeric}(ts)
}
\arguments{
\item{ts}{a [numeric] representing an epoch in seconds or a [POSIXt] date}
}
\value{
a [character] of the epoch in milliseconds
}
\description{
This function allows to convert epoch timestamp in seconds to an epoch in milliseconds
}
\examples{
Date2EpochMilli(as.numeric(Sys.time()))

}
