% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thingsboard_api.R
\name{ThingsboardApi_checkToken}
\alias{ThingsboardApi_checkToken}
\title{Check if the token is timeouted and refresh it if necessary}
\value{
\link{NULL}
}
\description{
Check if the token is timeouted and refresh it if necessary
}
