% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thingsboard_api.R
\name{ThingsboardApi_getTelemetry}
\alias{ThingsboardApi_getTelemetry}
\title{Fetch telemetry}
\arguments{
\item{...}{Parameters passed through method \link{ThingsboardApi_getValues}}

\item{endTs}{A \link{numeric} or a \link{POSIXct} representing respectively the epoch or the date of the end of data extraction period}
}
\value{
A \link{data.frame} with one row per data and 3 columns:
\itemize{
\item \code{key}: a \link{character} with the key
\item \code{ts}: a \link{POSIXct} with the timestamp of the data
\item \code{value}: a \link{numeric} with the value of the data
}
}
\description{
Fetch telemetry data for an entity. See \link{ThingsboardApi_getValues} for the arguments.

See: \url{https://thingsboard.io/docs/user-guide/telemetry/#get-telemetry-values}
}
