% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thingsboard_api.R
\name{ThingsboardApi_getToken}
\alias{ThingsboardApi_getToken}
\title{Get authorisation token from thingsboard server for a specific device}
\arguments{
\item{timeOut}{\link{numeric} number of second before token timeout (default field \code{tokenTimeOut})}
}
\value{
A \link{list} with keys 'token' and 'refreshtoken'
}
\description{
Get authorisation token from thingsboard server for a specific device
}
