% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unaccent.R
\name{Unaccent}
\alias{Unaccent}
\alias{UnaccentSmart}
\title{Remove accents from a string}
\usage{
Unaccent(text, from = "UTF-8")

UnaccentSmart(text, from = "UTF-8", alterfrom = "latin1")
}
\arguments{
\item{text}{the input string}

\item{from}{the initial charset (default: "UTF-8")}

\item{alterfrom}{the alternate charset (default: "Latin1")}
}
\value{
the string without accents
}
\description{
[Unaccent] remove accent from a string with a given charset, [UnaccentSmart]
do the same thing with an alternate charset in case of a string with
multiple mixed charsets.
}
\examples{
Unaccent("éàù")
UnaccentSmart("éàù")

}
\references{
From https://data.hypotheses.org/564
}
