% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thingsboard_api.R
\docType{class}
\name{ThingsboardApi-class}
\alias{ThingsboardApi-class}
\alias{ThingsboardApi}
\title{Thingboard API Class}
\description{
Thingboard API Class
}
\section{Fields}{

\describe{
\item{\code{url}}{URL of the thingsboard IoT platform.}

\item{\code{publicId}}{[character] the public ID of the device}

\item{\code{token}}{[character] the current token}

\item{\code{tokenTimeOut}}{A [numeric] contains the time out of a token in seconds (default 300)}

\item{\code{tokenEpiration}}{A [numeric] with the Epoch of the expiration date time of current token}
}}


\examples{
thinksboard_api = ThingsboardApi(url="http://scada.g-eau.net",
                                 publicId="299cedc0-f3e9-11e8-9dbf-cbc1e37c11e3")

}
