% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thingsboard_api.R
\name{ThingsboardApi_checkToken}
\alias{ThingsboardApi_checkToken}
\title{Check if the token is timeouted and refresh it if necessary}
\value{
\link{NULL}
}
\description{
This method is automatically call by each other methods of the class \link{ThingsboardApi}.
So, except for debugging purpose, it's not useful to call directly.
}
